+++ example/lib/dio_request_inspector_setting.dart
@@ -1,10 +1,9 @@
 import 'package:dio_request_inspector/dio_request_inspector.dart';
 
 final DioRequestInspector inspector = DioRequestInspector(
-  isDebugMode: true,
-  duration: const Duration(milliseconds: 500),
-  showFloating: true,
+  isInspectorEnabled: true,
   password: '123456', // remove this line if you don't need password
+  showSummary: false,
 );
 
 final void Function() toInspector = inspector.goToInspector;